Updated GitHub client
