+++ app/Http/Clients/GitHub/CachedClient.php
@@ -1,25 +1,24 @@
 <?php
 
-namespace App\GitHub;
+namespace App\Http\Clients\GitHub;
 
+use App\Http\Clients\GitHub\ClientInterface as GitHubClientInterface;
 use Carbon\CarbonInterval;
 use Illuminate\Support\Facades\Cache;
 use stdClass;
 
-class CachedClient extends Client
+class CachedClient implements GitHubClientInterface
 {
-    /**
-     * Get a cached release.
-     *
-     * @param string $owner The repository owner
-     * @param string $repo The repository name
-     */
+    public function __construct(
+        private GitHubClientInterface $github
+    ) {}
+
     public function latestRelease(string $owner, string $repo): stdClass
     {
         return Cache::remember(
             sprintf('latestRelease:%s:%s', $owner, $repo),
             CarbonInterval::minutes(rand(60, 120)),
-            fn (): stdClass => parent::latestRelease($owner, $repo)
+            fn (): stdClass => $this->github->latestRelease($owner, $repo)
         );
     }
 }
