+++ app/Http/Clients/GitHub/ClientInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Http\Clients\GitHub;
+
+use stdClass;
+
+interface ClientInterface
+{
+    /**
+     * Get the latest release for a repository.
+     *
+     * @param string $owner The repository owner
+     * @param string $repo The repository name
+     */
+    public function latestRelease(string $owner, string $repo): stdClass;
+}
