+++ app/Providers/AppServiceProvider.php
@@ -2,17 +2,29 @@
 
 namespace App\Providers;
 
-use App\GitHub\CachedClient as CachedGitHubClient;
-use App\GitHub\Client as GitHubClient;
+use App\Http\Clients\GitHub;
+use GuzzleHttp\Client as GuzzleHttpClient;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
 {
     public function register(): void
     {
-        $this->app->bind(GitHubClient::class, function ($app) {
-            return new CachedGitHubClient(config('services.github.token'));
+        $this->app->bind(GitHub\ClientInterface::class, function (): GitHub\Client {
+            return new GitHub\Client(
+                new GuzzleHttpClient([
+                    'base_uri' => Config::string('services.github.base_uri'),
+                    'connect_timeout' => 5,
+                    'headers' => [
+                        'Authorization' => sprintf('Token %s', Config::string('services.github.token')),
+                    ],
+                    'timeout' => 30,
+                ])
+            );
         });
+
+        $this->app->extend(GitHub\ClientInterface::class, fn (GitHub\ClientInterface $github): GitHub\ClientInterface => new GitHub\CachedClient($github));
     }
 
     public function boot(): void {}
